Move out the testing helpers to separate files
