+++ .gitignore
@@ -1,0 +1,2 @@
+# Exclude the IDE folders from the VCS
+./idea
